% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fleetbr}
\alias{fleetbr}
\title{Brazilian Vehicle Fleet Dataset}
\format{
\subsection{\code{fleetbr}}{

A \code{tibble} data frame with 92,664 and 5 columns:
\describe{
\item{uf}{Federative Unit - Brazilian States Acronyms}
\item{mes, ano}{Month and Year}
\item{modal}{Mode of transport of observation}
\item{frota}{Fleet size}
}
}
}
\source{
\url{https://www.gov.br/transportes/pt-br/assuntos/transito/conteudo-Senatran/estatisticas-frota-de-veiculos-senatran}
}
\usage{
fleetbr
}
\description{
The package includes a dataset from the road vehicle data provided
by the Brazilian Ministry of Transport.
}
\keyword{datasets}
