% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fleetbr}
\alias{fleetbr}
\title{Brazilian Ministry of Transportation Fleet data}
\format{
## `fleetbr`
A data.frame 28512 rows and 5 columns:
\describe{
   \item{\code{uf}}{Federative Unit - Brazilian State Acronyms}
   \item{\code{mes}}{Month of observation}
   \item{\code{ano}}{Year of observation}
   \item{\code{modal}}{Mode of transport}
   \item{\code{frota}}{Fleet size}
}
}
\source{
<https://www.gov.br/transportes/pt-br/assuntos/transito/conteudo-Senatran/estatisticas-frota-de-veiculos-senatran>
}
\usage{
fleetbr
}
\description{
Long format dataset from the road fleet data
provided by the Brazilian Ministry of Transportation
}
\keyword{datasets}
