% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fleetbr}
\alias{fleetbr}
\title{Brazilian Ministry of Transport Road Fleet data}
\format{
##`fleetbr`
A `tibble` data frame with 28,512 and 5 columns:
\describe{
  \item{uf}{Federative Unit - Brazilian States Acronyms}
  \item{mes, ano}{Month and Year}
  \item{modal}{Mode of transport of observation}
  \item{frota}{Fleet size}
}
}
\source{
<https://www.gov.br/transportes/pt-br/assuntos/transito/conteudo-Senatran/estatisticas-frota-de-veiculos-senatran>
}
\usage{
fleetbr
}
\description{
Long format dataset based on the road fleet databases
provided by the Brazilian Ministry of Transport. All variables
are encoded to Brazilian Portuguese.
}
\keyword{datasets}
