% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fleetbr}
\alias{fleetbr}
\title{Brazilian Ministry of Transport Road Fleet data}
\format{
\subsection{\code{fleetbr}}{

A \code{tibble} data frame with 78,408 and 5 columns:
\describe{
\item{uf}{Federative Unit - Brazilian States Acronyms}
\item{mes, ano}{Month and Year}
\item{modal}{Mode of transport of observation}
\item{frota}{Fleet size}
}
}
}
\source{
\url{https://www.gov.br/transportes/pt-br/assuntos/transito/conteudo-Senatran/estatisticas-frota-de-veiculos-senatran}
}
\usage{
fleetbr
}
\description{
This is a long format dataset based on the road fleet databases
provided by the Brazilian Ministry of Transport. All variables
are encoded to Brazilian Portuguese.
}
\keyword{datasets}
